requirements.txt
torchvision
func_timeout
scipy